+++ webcentral-backend-core/src/main/java/hu/bme/sch/bss/webcentral/core/DomainAuditModel.java
@@ -1,20 +1,14 @@
-package hu.bme.sch.bss.webcentral;
+package hu.bme.sch.bss.webcentral.core;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-
-import java.io.Serializable;
-import java.util.Date;
-
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.*;
+import java.io.Serializable;
+import java.util.Date;
 
 @JsonSerialize
 @MappedSuperclass
