+++ webcentral-backend-core/src/main/java/hu/bme/sch/bss/webcentral/core/dao/UserDao.java
@@ -1,6 +1,6 @@
-package hu.bme.sch.bss.webcentral.dao;
+package hu.bme.sch.bss.webcentral.core.dao;
 
-import hu.bme.sch.bss.webcentral.model.User;
+import hu.bme.sch.bss.webcentral.core.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
