+++ webcentral-backend-core/src/main/java/hu/bme/sch/bss/webcentral/core/model/User.java
@@ -1,8 +1,8 @@
-package hu.bme.sch.bss.webcentral.model;
+package hu.bme.sch.bss.webcentral.core.model;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import hu.bme.sch.bss.webcentral.DomainAuditModel;
+import hu.bme.sch.bss.webcentral.core.DomainAuditModel;
 
 import javax.persistence.*;
 import javax.validation.constraints.Email;
